removing comments and moving some code to Footer
