+++ src/i18n.js
@@ -5,19 +5,14 @@ import Backend from "i18next-http-backend";
 import { DateTime } from "luxon";
 
 i18n
-  // detect user language
-  // learn more: https://github.com/i18next/i18next-browser-languageDetector
   .use(Backend)
   .use(LanguageDetector)
-  // pass the i18n instance to react-i18next.
   .use(initReactI18next)
-  // init i18next
-  // for all options read: https://www.i18next.com/overview/configuration-options
   .init({
     debug: false,
     fallbackLng: "en",
     interpolation: {
-      escapeValue: false, // not needed for react as it escapes by default
+      escapeValue: false,
     },
     react: {
       useSuspense: false,
